fix: change primary key type to BigInteger

+++ src/app/entity/history.py
@@ -8,7 +8,7 @@ from datetime import datetime
 @dataclass
 class History(db.Model):
     __table_args__ = {'extend_existing': True}
-    id: int = db.Column(db.Integer, primary_key=True, autoincrement=True)
+    id: int = db.Column(db.BigInteger, primary_key=True, autoincrement=True)
     ip: str = db.Column(db.String(100))
     host: str = db.Column(db.String(50))
     status: bool = db.Column(db.Boolean)
